fix(scheduler): remove unnecessary cast in tool_count computation
`AgentLoopResult::iterations` is already `u32`, so the `as u32` cast
was redundant and triggered `clippy::unnecessary_cast` with `-D warnings`.

+++ crates/librefang-kernel/src/kernel.rs
@@ -2891,7 +2891,7 @@ system_prompt = "You are a helpful assistant."
                 // Record token usage for quota tracking
                 self.scheduler.record_usage(agent_id, &result.total_usage);
                 // Record tool calls for rate limiting
-                let tool_count = result.iterations.saturating_sub(1) as u32;
+                let tool_count = result.iterations.saturating_sub(1);
                 self.scheduler.record_tool_calls(agent_id, tool_count);
 
                 // Update last active time
@@ -3438,7 +3438,7 @@ system_prompt = "You are a helpful assistant."
                         .scheduler
                         .record_usage(agent_id, &result.total_usage);
                     // Record tool calls for rate limiting
-                    let tool_count = result.iterations.saturating_sub(1) as u32;
+                    let tool_count = result.iterations.saturating_sub(1);
                     kernel_clone
                         .scheduler
                         .record_tool_calls(agent_id, tool_count);
